+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ### vNEXT
 
+
+### 1.7.0
+- Add support for network interfaces that return observables [PR #1840](https://github.com/apollographql/apollo-client/pull/1840)
+
+### 1.6.1
+- Pin @types/node to 8.0.2 to avoid breaking type update
+
+### 1.6.0
+- the `@connection(key: ...)` directive can now be used to specify the key to use
+for the Apollo store and is removed by default when sending queries to the server [PR #1801](https://github.com/apollographql/apollo-client/pull/1801)
+
 ### 1.5.0
 - `batchInterval` now has a default value of 10 ms [PR #1793](https://github.com/apollographql/apollo-client/pull/1793)
 - Added `batchMax` to allow you to limit the amount of queries in one batch. [PR #1659](https://github.com/apollographql/apollo-client/pull/1659)